Fix double-borrow in UIBindings macro. Fix (at least some of) the incorrect types in Components' UIBindings.
Stabilizers changes:
- Convert hstab_sel and vstab_sel from i16 to usize
- Change UIBindings from number to select type with source attributes
- Update all methods to work with usize instead of i16
- Remove >= 0 checks (usize is always >= 0)
- Remove "as usize" casts when indexing
- Update JSON serialization/deserialization to use usize

ControlSurfaces changes:
- Fix drag_sel TypeScript handling for check_list bindings
- Update createDragInducersSection to use flat array structure
- Update updateValues() to handle check_list binding correctly
- check_list returns array of {name, enabled, selected} objects

+++ flyingcircusrust/src/stabilizers/json.rs
@@ -5,13 +5,13 @@ use serde_json::json;
 impl JSSerializable for Stabilizers {
     fn from_json(&mut self, json: &serde_json::Value, _version: f32) {
         if let Some(val) = json["hstab_sel"].as_i64() {
-            self.hstab_sel = val as i16;
+            self.hstab_sel = val as usize;
         }
         if let Some(val) = json["hstab_count"].as_i64() {
             self.hstab_count = val as i16;
         }
         if let Some(val) = json["vstab_sel"].as_i64() {
-            self.vstab_sel = val as i16;
+            self.vstab_sel = val as usize;
         }
         if let Some(val) = json["vstab_count"].as_i64() {
             self.vstab_count = val as i16;
